Homework - Lesson_3 level_7 - some packages created and some classes moved here
