+++ src/main/java/students/student_svetlana_seda/lesson_3/level_7/task_29/Cat.java
@@ -1,4 +1,4 @@
-package students.student_svetlana_seda.lesson_3.level_7;
+package students.student_svetlana_seda.lesson_3.level_7.task_29;
 
 public class Cat {
     String name;
