+++ src/main/java/students/student_svetlana_seda/lesson_3/level_7/task_29/GirlDemo.java
@@ -1,4 +1,6 @@
-package students.student_svetlana_seda.lesson_3.level_7;
+package students.student_svetlana_seda.lesson_3.level_7.task_29;
+
+import students.student_svetlana_seda.lesson_3.level_7.task_29.Girl;
 
 public class GirlDemo {
