+++ src/main/java/students/student_svetlana_seda/lesson_3/level_7/task_31/ProductDemo.java
@@ -1,4 +1,6 @@
-package students.student_svetlana_seda.lesson_3.level_7;
+package students.student_svetlana_seda.lesson_3.level_7.task_31;
+
+import students.student_svetlana_seda.lesson_3.level_7.task_31.Product;
 
 public class ProductDemo {
